pseudoarreglo de rutas
