--- src/controllers/carritoController.js
@@ -1,7 +1,0 @@
-const carritoController = {
-  carrito: (req, res) => {
-    return res.render("carrito");
-  }
-};
-
-module.exports = carritoController;
