--- src/controllers/formularioController.js
@@ -1,7 +1,0 @@
-const formularioController = {
-    formulario: (req, res) => {
-      return res.render("formulario");
-    }
-  };
-  
-  module.exports = formularioControllerController;
