--- src/controllers/homeController.js
@@ -1,7 +1,0 @@
-const homeController = {
-  home: (req, res) => {
-    return res.render("home");
-  }
-};
-
-module.exports = homeController;
