+++ src/controllers/indexController.js
@@ -1,0 +1,7 @@
+const indexController = {
+  index: (req, res) => {
+    return res.render("home");
+  },
+};
+
+module.exports = indexController;
