--- src/controllers/loginController.js
@@ -1,7 +1,0 @@
-const loginController = {
-  login: (req, res) => {
-    return res.render("login");
-  }
-};
-
-module.exports = loginController;
