--- src/controllers/productDetailController.js
@@ -1,7 +1,0 @@
-const productDetailController = {
-  productDetail: (req, res) => {
-    return res.render("productDetail");
-  }
-};
-
-module.exports = productDetailController;
