--- src/controllers/registerController.js
@@ -1,7 +1,0 @@
-const registerController = {
-  register: (req, res) => {
-    return res.render("register");
-  }
-};
-
-module.exports = registerController;
