--- src/routers/formulario.routes.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const formularioController = require("../controllers/formularioController");
-
-router.get("/", formularioController.formulario);
-
-module.exports = router;
