--- src/routers/home.routes.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const homeController = require("../controllers/homeController");
-
-router.get("/", homeController.home);
-
-module.exports = router;
