+++ src/routers/index.routes.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const indexController = require("../controllers/indexController");
+
+router.get("/", indexController.index);
+
+module.exports = indexController;
