--- src/routers/login.routes.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const loginController = require("../controllers/loginController");
-
-router.get("/", loginController.login);
-
-module.exports = router;
