--- src/routers/productDetail.routes.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const productDetailController = require("../controllers/productDetailController");
-
-router.get("/", productDetailController.productDetail);
-
-module.exports = router;
