--- src/routers/register.routes.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const registerController = require("../controllers/registerController");
-
-router.get("/", registerController.register);
-
-module.exports = router;
