+++ src/routers/user.routes.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+const userController = require("../controllers/userController");
+
+router.get("/", userController.login);
+router.get("/", userController.register);
+
+module.exports = userController;
