+++ src/views/home.ejs
@@ -1,9 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <link rel="website icon" type="png"
-    href="images/brandmark@2x-8.png">
-    
+    <link rel="website icon" type="png" href="images/brandmark@2x-8.png" />
+
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Bijou Boutique</title>
@@ -66,26 +65,6 @@
         </div>
       </section>
 
-      <section class="item-especial">
-        <div class="flexible">
-          <div class="itemFlex1">
-            <img src="/images/reloj.jpg" alt="imgFlex" />
-          </div>
-          <div class="itemFlex2">
-            <h4>Modelo iconico</h4>
-            <h3>Modelo Exclusivo: Alpine Eagle</h3>
-            <div class="separator"></div>
-            <p>
-              Descubre la sofisticación en cada momento con nuestro reloj
-              elegante. Con un diseño moderno y funcionalidades avanzadas, este
-              reloj es la elección perfecta para ocasiones especiales y tu día a
-              día. Conoce la combinación única de estilo y precisión que te
-              ofrece este exclusivo reloj de Bijou Boutique.
-            </p>
-          </div>
-        </div>
-      </section>
-
       <section class="carousel-section">
         <h2>Explora Nuestros Productos Destacados</h2>
         <div class="slick-carousel">
@@ -97,9 +76,9 @@
         </div>
       </section>
 
-      <section class="news-section">
-        <h2>Últimas Novedades</h2>
+      <h2 class="ultimas-novedades">Últimas Novedades</h2>
 
+      <section class="news-section">
         <article class="news-item">
           <img src="/images/novedades1.png" alt="Novedad 1" />
           <h3>Gran Apertura de la Tienda</h3>
@@ -119,6 +98,26 @@
           </p>
           <a href="#">Leer más</a>
         </article>
+      </section>
+
+      <section class="item-especial">
+        <div class="flexible">
+          <div class="itemFlex1">
+            <img src="/images/reloj.jpg" alt="imgFlex" />
+          </div>
+          <div class="itemFlex2">
+            <h1>Modelo iconico</h1>
+            <h2>Alpine Eagle</h2>
+            <div class="separator"></div>
+            <p>
+              Descubre la sofisticación en cada momento con nuestro reloj
+              elegante. Con un diseño moderno y funcionalidades avanzadas, este
+              reloj es la elección perfecta para ocasiones especiales y tu día a
+              día. Conoce la combinación única de estilo y precisión que te
+              ofrece este exclusivo reloj de Bijou Boutique.
+            </p>
+          </div>
+        </div>
       </section>
 
       <section class="products-section">
