+++ src/views/login.ejs
@@ -1,32 +1,40 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-
-  <link rel="website icon" type="png" 
-    href="images/brandmark@2x-8.png">
-    
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="/css/login.css">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="/css/login.css" />
     <title>Bijou Boutique</title>
-</head>
-<body>
+  </head>
+  <body>
     <%- include("partials/header.ejs") %>
-    
-    <div class="login-container">
-        <div class="login-form">
-          <img src="/images/logo3.png" alt="" class="logo">
-          <form>
-            <label for="email" class="label">Email</label><br>
-            <input type="email" id="email" name="email" placeholder="Email" class="input input-email"><br>
-    
-            <label for="password" class="label">Password:</label><br>
-            <input type="password" id="password" name="passwsord" placeholder="*********" class="input input-password"><br>
-          </form>
-          <button class="secondary-button signup-button">Registrarse</button>
-        </div>
+
+    <main class="container-register">
+      <div class="form">
+        <form>
+          <h2 class="title">Iniciar sesión</h2>
+          <input
+            class="controls"
+            id="user"
+            type="text"
+            placeholder="Nombre de usuario"
+          />
+          <input
+            class="controls"
+            id="pass"
+            type="password"
+            placeholder="Contraseña"
+          />
+
+          <a href="/register" class="register-link"
+            ><h5 class="title">¿No tienes cuenta aún? Click aquí!</h5></a
+          >
+
+          <input class="submit" type="submit" />
+        </form>
       </div>
+    </main>
 
     <%- include("partials/footer.ejs") %>
-</body>
+  </body>
 </html>
