+++ src/views/partials/header.ejs
@@ -1,9 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-    <link rel="website icon" type="png"
-    href="images/brandmark@2x-8.png">
-    
+    <link rel="website icon" type="png" href="images/brandmark@2x-8.png" />
+
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Bijou Boutique</title>
@@ -19,7 +18,9 @@
   <body>
     <header>
       <div class="logo-container">
-        <a href="/"><img src="/images/bijou-boutique.png" alt="Logo de la tienda" /></a>
+        <a href="/"
+          ><img src="/images/bijou-boutique.png" alt="Logo de la tienda"
+        /></a>
       </div>
 
       <div class="icons-container">
@@ -31,5 +32,5 @@
         <!-- Icono de Artículos Favoritos -->
       </div>
     </header>
-
-    
+  </body>
+</html>
