+++ src/views/register.ejs
@@ -74,18 +74,6 @@
             type="reset"
             onclick="alert('Datos reestablecidos')"
           />
-          <div class="content">
-            <h1>Registro de Usuario</h1>
-            <form>
-              <!-- Tus campos de registro aquí -->
-              <button type="submit" style="background-color: #003366">
-                Registrar
-              </button>
-              <button type="reset" style="background-color: #003366">
-                Restablecer
-              </button>
-            </form>
-          </div>
         </form>
       </div>
     </main>
